requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=2.9.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
# spacy>=3.7.0  # Optional - incompatible with Python 3.14. Entity extraction uses regex fallback.
pypdf2>=3.0.1
numpy>=1.24.0
